refactor: split admin users handlers from deposits
